+++ tests/integration/components/taxonomy-top-list-test.js
@@ -1,4 +1,5 @@
-import { A } from '@ember/array';
+import EmberObject from '@ember/object';
+
 import { moduleForComponent, test } from 'ember-qunit';
 import hbs from 'htmlbars-inline-precompile';
 
@@ -7,7 +8,10 @@ moduleForComponent('taxonomy-top-list', 'Integration | Component | taxonomy top 
 });
 
 test('it renders', function(assert) {
-    this.set('list', A([{ text: 'b' }, { text: 'a' }]));
+    this.set('list', [
+        EmberObject.create({ text: 'b' }),
+        EmberObject.create({ text: 'a' }),
+    ]);
     this.render(hbs`{{taxonomy-top-list list=list}}`);
 
     // Should be sorted as 'ab', not 'ba'